requirements.txt

ultralytics>=8.2.0
torch>=2.1.0
torchvision>=0.16.0
numpy
opencv-python
pandas
pyyaml
python-dotenv
tqdm
scikit-learn
kaggle
matplotlib
